refactor: moved GC vehicle-related files to new folder
