+++ src/main/java/dev/galacticraft/mod/client/render/entity/BuggyRenderer.java
@@ -30,7 +30,7 @@ import dev.galacticraft.mod.client.model.GCBakedModel;
 import dev.galacticraft.mod.client.model.GCModelLoader;
 import dev.galacticraft.mod.client.model.GCModelState;
 import dev.galacticraft.mod.client.model.GCSheets;
-import dev.galacticraft.mod.content.entity.Buggy;
+import dev.galacticraft.mod.content.entity.vehicle.Buggy;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.entity.EntityRenderer;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
