+++ src/main/java/dev/galacticraft/mod/content/GCEntityTypes.java
@@ -27,8 +27,9 @@ import dev.galacticraft.mod.Constant;
 import dev.galacticraft.mod.Constant.Entity;
 import dev.galacticraft.mod.content.entity.*;
 import dev.galacticraft.mod.content.entity.boss.SkeletonBoss;
-import dev.galacticraft.mod.content.entity.orbital.RocketEntity;
-import dev.galacticraft.mod.content.entity.orbital.lander.LanderEntity;
+import dev.galacticraft.mod.content.entity.vehicle.Buggy;
+import dev.galacticraft.mod.content.entity.vehicle.RocketEntity;
+import dev.galacticraft.mod.content.entity.vehicle.LanderEntity;
 import net.fabricmc.fabric.api.object.builder.v1.entity.FabricDefaultAttributeRegistry;
 import net.fabricmc.fabric.api.object.builder.v1.entity.FabricEntityTypeBuilder;
 import net.minecraft.core.registries.BuiltInRegistries;
