+++ src/main/java/dev/galacticraft/mod/content/entity/vehicle/LanderEntity.java
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-package dev.galacticraft.mod.content.entity.orbital.lander;
+package dev.galacticraft.mod.content.entity.vehicle;
 
 import com.mojang.datafixers.util.Pair;
 import dev.galacticraft.api.entity.IgnoreShift;
