+++ src/main/java/dev/galacticraft/mod/network/GCServerPacketReceivers.java
@@ -27,7 +27,7 @@ import dev.galacticraft.api.registry.AddonRegistries;
 import dev.galacticraft.api.universe.celestialbody.CelestialBody;
 import dev.galacticraft.impl.universe.celestialbody.type.SatelliteType;
 import dev.galacticraft.mod.Constant;
-import dev.galacticraft.mod.content.entity.orbital.RocketEntity;
+import dev.galacticraft.mod.content.entity.vehicle.RocketEntity;
 import dev.galacticraft.mod.events.GCEventHandlers;
 import dev.galacticraft.mod.network.packets.*;
 import dev.galacticraft.mod.screen.GCPlayerInventoryMenu;
